+++ packages/editor-common/web/src/tiptapTypes.ts
@@ -1,6 +1,7 @@
 import { Node as ProseMirrorNode } from 'prosemirror-model';
 import { FC } from 'react';
 import { Editor } from '@tiptap/react';
+import { LegacyEditorPluginConfig } from 'wix-rich-content-common';
 export * from './Utils/tiptap/types';
 
 export interface PluginProps {
@@ -8,6 +9,7 @@ export interface PluginProps {
     isMobile: boolean;
     theme: string;
     t: (key: string) => string;
+    config: LegacyEditorPluginConfig;
   };
   // eslint-disable-next-line
   componentData: any;