+++ packages/editor-common/web/src/Components/LinkComponents/LinkPanelDropdown.jsx
@@ -1,6 +1,7 @@
 import React, { Component, PureComponent, Suspense, lazy } from 'react';
 import PropTypes from 'prop-types';
 import styles from '../../../statics/styles/link-panel.scss';
+import { TextInput } from 'wix-rich-content-ui-components';
 import { mergeStyles } from 'wix-rich-content-common';
 import { isUndefined } from 'lodash';
 
@@ -65,6 +66,8 @@ export class LinkPanelDropdown extends Component {
     fallbackChanged: false,
   };
 
+  textInput = React.createRef();
+
   styles = mergeStyles({ styles, theme: this.props.theme });
 
   handleDropDownStateChange = changes => {
@@ -84,20 +87,32 @@ export class LinkPanelDropdown extends Component {
     }
   };
 
+  componentDidMount() {
+    // eslint-disable-next-line react/prop-types
+    this.textInput.current?.focus();
+    if (!this.state.fallbackChanged) {
+      this.textInput.current?.select(); //select the link in case of edit
+    } else {
+      this.textInput.current.selectionStart = this.textInput.current?.value.length;
+      this.textInput.current.selectionEnd = this.textInput.current?.value.length;
+    }
+  }
+
   render() {
     const { itemToString, formatMenuItem, itemHeight, textInputProps, value } = this.props;
-    const { selectedItem, items, fallbackChanged } = this.state;
+    const { selectedItem, items } = this.state;
+    const fallBackOnchange = value => {
+      this.props.onChange(value);
+      this.setState({ fallbackChanged: true });
+    };
     return (
       <Suspense
         fallback={
-          <Input
+          <TextInput
             {...textInputProps}
+            inputRef={this.textInput}
             value={value}
-            onChange={e => {
-              this.props.onChange(e.target.value);
-              this.setState({ fallbackChanged: true });
-            }}
-            selectText
+            onChange={fallBackOnchange}
           />
         }
       >
@@ -115,34 +130,42 @@ export class LinkPanelDropdown extends Component {
             isOpen,
             highlightedIndex,
             inputValue,
-          }) => (
-            <div>
-              {/*<label {...getLabelProps()}>Enter a fruit</label>*/}
-              <Input {...getInputProps({ ...textInputProps })} selectText={!fallbackChanged} />
-              {(isOpen || this.props.isOpen) && List && (
-                <Suspense fallback={<div>Loading...</div>}>
-                  <List
-                    className={styles.linkPanel_dropdownList}
-                    style={{ borderTop: '0', position: 'absolute' }}
-                    height={Math.min(items.length * itemHeight + 1, 200)}
-                    itemCount={items.length}
-                    itemSize={itemHeight}
-                    itemData={{
-                      items,
-                      getItemProps,
-                      highlightedIndex,
-                      selectedItem,
-                      formatMenuItem,
-                      inputValue,
-                    }}
-                    {...getMenuProps()}
-                  >
-                    {ItemRenderer}
-                  </List>
-                </Suspense>
-              )}
-            </div>
-          )}
+          }) => {
+            const textInputProps = getInputProps({
+              ...textInputProps,
+            });
+            return (
+              <div>
+                {/*<label {...getLabelProps()}>Enter a fruit</label>*/}
+                <TextInput
+                  {...getInputProps({ ...textInputProps })}
+                  onChange={value => textInputProps.onChange({ target: { value } })}
+                  inputRef={this.textInput}
+                />
+                {(isOpen || this.props.isOpen) && List && (
+                  <Suspense fallback={<div>Loading...</div>}>
+                    <List
+                      className={styles.linkPanel_dropdownList}
+                      height={Math.min(items.length * itemHeight + 1, 200)}
+                      itemCount={items.length}
+                      itemSize={itemHeight}
+                      itemData={{
+                        items,
+                        getItemProps,
+                        highlightedIndex,
+                        selectedItem,
+                        formatMenuItem,
+                        inputValue,
+                      }}
+                      {...getMenuProps()}
+                    >
+                      {ItemRenderer}
+                    </List>
+                  </Suspense>
+                )}
+              </div>
+            );
+          }}
         </Downshift>
       </Suspense>
     );
@@ -159,26 +182,4 @@ export class LinkPanelDropdown extends Component {
     textInputProps: PropTypes.object,
     isOpen: PropTypes.bool,
   };
-}
-
-class Input extends Component {
-  textInput = React.createRef();
-
-  componentDidMount() {
-    // eslint-disable-next-line react/prop-types
-    const { selectText } = this.props;
-    this.textInput.current.focus();
-    if (selectText) {
-      this.textInput.current.select(); //select the link in case of edit
-    } else {
-      this.textInput.current.selectionStart = this.textInput.current.value.length;
-      this.textInput.current.selectionEnd = this.textInput.current.value.length;
-    }
-  }
-
-  render() {
-    // eslint-disable-next-line react/prop-types, no-unused-vars
-    const { selectText, ...inputProps } = this.props;
-    return <input {...inputProps} ref={this.textInput} />;
-  }
 }