+++ packages/ricos-editor/web/src/RicosEditor.tsx
@@ -267,8 +267,6 @@ export class RicosEditor extends Component<RicosEditorProps, State> {
 
   getT = () => this.editor.getT();
 
-  removeToolbarFocus = () => this.editor.removeToolbarFocus();
-
   renderToolbarPortal(Toolbar) {
     return (
       <StaticToolbarPortal
@@ -401,6 +399,8 @@ export class RicosEditor extends Component<RicosEditorProps, State> {
     );
     const { localeData } = this.state;
     const { locale, localeResource } = localeData;
+    // TODO: not a complete config
+    const config = plugins?.reduce((prev, curr) => ({ ...prev, [curr.type]: curr.config }), {});
     return (
       <Fragment>
         {tiptapToolbar && this.renderToolbarPortal(tiptapToolbar)}
@@ -411,6 +411,7 @@ export class RicosEditor extends Component<RicosEditorProps, State> {
                 <RicosTiptapEditor
                   extensions={plugins}
                   content={initalContent}
+                  config={config}
                   t={t}
                   onLoad={editor => {
                     const richContentAdapter = new RichContentAdapter(editor);