+++ packages/tiptap-editor/web/src/components/RicosTiptapEditor/RicosTiptapEditor.tsx
@@ -4,8 +4,6 @@ import { EditorContent, Editor } from '@tiptap/react';
 
 import { RicosExtensionManager } from '../../ricos-extensions-manager';
 import { tiptapExtensions as coreExtensions } from '../../tiptap-extensions';
-import { createImageConfig } from '../../extensions/extension-image';
-import { createSpoilerConfig } from '../../extensions/extension-spoiler';
 import { createDraftConfig } from '../../extensions/extension-draft';
 import { createHistoryConfig } from '../../extensions/extension-history';
 import { createStylesConfig } from '../../extensions/extension-styles';
@@ -16,8 +14,6 @@ function useForceUpdate() {
 
   return () => setValue(value => value + 1);
 }
-const imageConfig = createImageConfig();
-const spoilerConfig = createSpoilerConfig();
 const draftConfig = createDraftConfig();
 const historyConfig = createHistoryConfig();
 const stylesConfig = createStylesConfig();
@@ -27,7 +23,7 @@ export const RicosTiptapEditor = ({ content, extensions = [], onLoad, ...context
   const [nodeViewsHOCs, setNodeViewsHOCs] = useState([]);
   const forceUpdate = useForceUpdate();
 
-  const extensionsConfigs = [imageConfig, spoilerConfig, draftConfig, historyConfig, stylesConfig];
+  const extensionsConfigs = [draftConfig, historyConfig, stylesConfig];
   useEffect(() => {
     const ricosExtensionsConfigs = createRicosExtensionsConfigs(extensions);
     const allConfigs = [...ricosExtensionsConfigs, ...extensionsConfigs];