+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 
 - `tiptap`
   - [#2815](https://github.com/wix/ricos/pull/2815) external `addNodeViewHOC` for generic extensions
+  - [#2825](https://github.com/wix/ricos/pull/2825) tiptap focus extensions
 
 ### :bug: Bug Fix